split deposits, deposit_data, deposit_parameters, casper_slashings, and slashable_vote_data out
